+++ cmd/frontend/internal/httpapi/internal.go
@@ -31,7 +31,7 @@ func serveReposGetByName(w http.ResponseWriter, r *http.Request) error {
 		return err
 	}
 	w.WriteHeader(http.StatusOK)
-	w.Write(data)
+	_, _ = w.Write(data)
 	return nil
 }
 
@@ -50,7 +50,7 @@ func servePhabricatorRepoCreate(w http.ResponseWriter, r *http.Request) error {
 		return err
 	}
 	w.WriteHeader(http.StatusOK)
-	w.Write(data)
+	_, _ = w.Write(data)
 	return nil
 }
 
@@ -260,7 +260,7 @@ func (h *reposListServer) serveIndex(w http.ResponseWriter, r *http.Request) err
 		}
 	} else {
 		trueP := true
-		res, err := h.Repos.List(r.Context(), db.ReposListOptions{Index: &trueP, Enabled: true})
+		res, err := h.Repos.List(r.Context(), db.ReposListOptions{Index: &trueP})
 		if err != nil {
 			return errors.Wrap(err, "listing repos")
 		}
@@ -362,7 +362,7 @@ func serveSavedQueriesSetInfo(w http.ResponseWriter, r *http.Request) error {
 		return errors.Wrap(err, "SavedQueries.Set")
 	}
 	w.WriteHeader(http.StatusOK)
-	w.Write([]byte("OK"))
+	_, _ = w.Write([]byte("OK"))
 	return nil
 }
 
@@ -377,7 +377,7 @@ func serveSavedQueriesDeleteInfo(w http.ResponseWriter, r *http.Request) error {
 		return errors.Wrap(err, "SavedQueries.Delete")
 	}
 	w.WriteHeader(http.StatusOK)
-	w.Write([]byte("OK"))
+	_, _ = w.Write([]byte("OK"))
 	return nil
 }
 
@@ -500,7 +500,7 @@ func serveGitResolveRevision(w http.ResponseWriter, r *http.Request) error {
 	}
 
 	w.WriteHeader(http.StatusOK)
-	w.Write([]byte(commitID))
+	_, _ = w.Write([]byte(commitID))
 	return nil
 }