+++ lsif/src/shared/database/postgres.ts
@@ -16,7 +16,7 @@ import { TlsOptions } from 'tls'
  * version prior to making use of the DB (which the frontend may still be
  * migrating).
  */
-const MINIMUM_MIGRATION_VERSION = 1528395634
+const MINIMUM_MIGRATION_VERSION = 1528395637
 
 /**
  * How many times to try to check the current database migration version on startup.