+++ lsif/src/tests/integration/backend/xrepo.test.ts
@@ -4,11 +4,21 @@ describe('Backend', () => {
     const ctx = new util.BackendTestContext()
     const commit = util.createCommit()
 
+    const ids = {
+        a: 100,
+        b1: 101,
+        b2: 103,
+        b3: 104,
+        c1: 105,
+        c2: 106,
+        c3: 107,
+    }
+
     beforeAll(async () => {
         await ctx.init()
         await Promise.all(
-            ['a', 'b1', 'b2', 'b3', 'c1', 'c2', 'c3'].map(r =>
-                ctx.convertTestData(r, commit, '', `xrepo/data/${r}.lsif.gz`)
+            Object.entries(ids).map(([repositoryName, repositoryId]) =>
+                ctx.convertTestData(repositoryId, repositoryName, commit, '', `xrepo/data/${repositoryName}.lsif.gz`)
             )
         )
     })
@@ -22,12 +32,12 @@ describe('Backend', () => {
             fail('failed beforeAll')
         }
 
-        const definitions = await ctx.backend.definitions('a', commit, 'src/index.ts', {
+        const definitions = await ctx.backend.definitions(ids.a, 'a', commit, 'src/index.ts', {
             line: 11,
             character: 18,
         })
         expect(definitions?.map(util.mapLocation)).toEqual([
-            util.createLocation('a', commit, 'src/index.ts', 0, 16, 0, 19),
+            util.createLocation(ids.a, commit, 'src/index.ts', 0, 16, 0, 19),
         ])
     })
 
@@ -36,12 +46,12 @@ describe('Backend', () => {
             fail('failed beforeAll')
         }
 
-        const definitions = await ctx.backend.definitions('b1', commit, 'src/index.ts', {
+        const definitions = await ctx.backend.definitions(ids.b1, 'b1', commit, 'src/index.ts', {
             line: 3,
             character: 12,
         })
         expect(definitions?.map(util.mapLocation)).toEqual([
-            util.createLocation('a', commit, 'src/index.ts', 0, 16, 0, 19),
+            util.createLocation(ids.a, commit, 'src/index.ts', 0, 16, 0, 19),
         ])
     })
 
@@ -50,12 +60,12 @@ describe('Backend', () => {
             fail('failed beforeAll')
         }
 
-        const definitions = await ctx.backend.definitions('b1', commit, 'src/index.ts', {
+        const definitions = await ctx.backend.definitions(ids.b1, 'b1', commit, 'src/index.ts', {
             line: 3,
             character: 16,
         })
         expect(definitions?.map(util.mapLocation)).toEqual([
-            util.createLocation('a', commit, 'src/index.ts', 4, 16, 4, 19),
+            util.createLocation(ids.a, commit, 'src/index.ts', 4, 16, 4, 19),
         ])
     })
 
@@ -66,23 +76,23 @@ describe('Backend', () => {
 
         const { locations } = util.filterNodeModules(
             util.mapLocations(
-                (await ctx.backend.references('a', util.createCommit(0), 'src/index.ts', {
+                (await ctx.backend.references(ids.a, 'a', util.createCommit(0), 'src/index.ts', {
                     line: 4,
                     character: 19,
                 })) || { locations: [] }
             )
         )
 
-        expect(locations).toContainEqual(util.createLocation('a', commit, 'src/index.ts', 4, 16, 4, 19)) // def
-        expect(locations).toContainEqual(util.createLocation('b1', commit, 'src/index.ts', 0, 14, 0, 17)) // import
-        expect(locations).toContainEqual(util.createLocation('b1', commit, 'src/index.ts', 3, 15, 3, 18)) // 1st use
-        expect(locations).toContainEqual(util.createLocation('b1', commit, 'src/index.ts', 3, 26, 3, 29)) // 2nd use
-        expect(locations).toContainEqual(util.createLocation('b2', commit, 'src/index.ts', 0, 14, 0, 17)) // import
-        expect(locations).toContainEqual(util.createLocation('b2', commit, 'src/index.ts', 3, 15, 3, 18)) // 1st use
-        expect(locations).toContainEqual(util.createLocation('b2', commit, 'src/index.ts', 3, 26, 3, 29)) // 2nd use
-        expect(locations).toContainEqual(util.createLocation('b3', commit, 'src/index.ts', 0, 14, 0, 17)) // import
-        expect(locations).toContainEqual(util.createLocation('b3', commit, 'src/index.ts', 3, 15, 3, 18)) // 1st use
-        expect(locations).toContainEqual(util.createLocation('b3', commit, 'src/index.ts', 3, 26, 3, 29)) // 2nd use
+        expect(locations).toContainEqual(util.createLocation(ids.a, commit, 'src/index.ts', 4, 16, 4, 19)) // def
+        expect(locations).toContainEqual(util.createLocation(ids.b1, commit, 'src/index.ts', 0, 14, 0, 17)) // import
+        expect(locations).toContainEqual(util.createLocation(ids.b1, commit, 'src/index.ts', 3, 15, 3, 18)) // 1st use
+        expect(locations).toContainEqual(util.createLocation(ids.b1, commit, 'src/index.ts', 3, 26, 3, 29)) // 2nd use
+        expect(locations).toContainEqual(util.createLocation(ids.b2, commit, 'src/index.ts', 0, 14, 0, 17)) // import
+        expect(locations).toContainEqual(util.createLocation(ids.b2, commit, 'src/index.ts', 3, 15, 3, 18)) // 1st use
+        expect(locations).toContainEqual(util.createLocation(ids.b2, commit, 'src/index.ts', 3, 26, 3, 29)) // 2nd use
+        expect(locations).toContainEqual(util.createLocation(ids.b3, commit, 'src/index.ts', 0, 14, 0, 17)) // import
+        expect(locations).toContainEqual(util.createLocation(ids.b3, commit, 'src/index.ts', 3, 15, 3, 18)) // 1st use
+        expect(locations).toContainEqual(util.createLocation(ids.b3, commit, 'src/index.ts', 3, 26, 3, 29)) // 2nd use
 
         // Ensure no additional references
         expect(locations?.length).toEqual(10)
@@ -95,23 +105,23 @@ describe('Backend', () => {
 
         const { locations } = util.filterNodeModules(
             util.mapLocations(
-                (await ctx.backend.references('b1', util.createCommit(0), 'src/index.ts', {
+                (await ctx.backend.references(ids.b1, 'b1', util.createCommit(0), 'src/index.ts', {
                     line: 3,
                     character: 16,
                 })) || { locations: [] }
             )
         )
 
-        expect(locations).toContainEqual(util.createLocation('a', commit, 'src/index.ts', 4, 16, 4, 19)) // def
-        expect(locations).toContainEqual(util.createLocation('b1', commit, 'src/index.ts', 0, 14, 0, 17)) // import
-        expect(locations).toContainEqual(util.createLocation('b1', commit, 'src/index.ts', 3, 15, 3, 18)) // 1st use
-        expect(locations).toContainEqual(util.createLocation('b1', commit, 'src/index.ts', 3, 26, 3, 29)) // 2nd use
-        expect(locations).toContainEqual(util.createLocation('b2', commit, 'src/index.ts', 0, 14, 0, 17)) // import
-        expect(locations).toContainEqual(util.createLocation('b2', commit, 'src/index.ts', 3, 15, 3, 18)) // 1st use
-        expect(locations).toContainEqual(util.createLocation('b2', commit, 'src/index.ts', 3, 26, 3, 29)) // 2nd use
-        expect(locations).toContainEqual(util.createLocation('b3', commit, 'src/index.ts', 0, 14, 0, 17)) // import
-        expect(locations).toContainEqual(util.createLocation('b3', commit, 'src/index.ts', 3, 15, 3, 18)) // 1st use
-        expect(locations).toContainEqual(util.createLocation('b3', commit, 'src/index.ts', 3, 26, 3, 29)) // 2nd use
+        expect(locations).toContainEqual(util.createLocation(ids.a, commit, 'src/index.ts', 4, 16, 4, 19)) // def
+        expect(locations).toContainEqual(util.createLocation(ids.b1, commit, 'src/index.ts', 0, 14, 0, 17)) // import
+        expect(locations).toContainEqual(util.createLocation(ids.b1, commit, 'src/index.ts', 3, 15, 3, 18)) // 1st use
+        expect(locations).toContainEqual(util.createLocation(ids.b1, commit, 'src/index.ts', 3, 26, 3, 29)) // 2nd use
+        expect(locations).toContainEqual(util.createLocation(ids.b2, commit, 'src/index.ts', 0, 14, 0, 17)) // import
+        expect(locations).toContainEqual(util.createLocation(ids.b2, commit, 'src/index.ts', 3, 15, 3, 18)) // 1st use
+        expect(locations).toContainEqual(util.createLocation(ids.b2, commit, 'src/index.ts', 3, 26, 3, 29)) // 2nd use
+        expect(locations).toContainEqual(util.createLocation(ids.b3, commit, 'src/index.ts', 0, 14, 0, 17)) // import
+        expect(locations).toContainEqual(util.createLocation(ids.b3, commit, 'src/index.ts', 3, 15, 3, 18)) // 1st use
+        expect(locations).toContainEqual(util.createLocation(ids.b3, commit, 'src/index.ts', 3, 26, 3, 29)) // 2nd use
 
         // Ensure no additional references
         expect(locations?.length).toEqual(10)
@@ -124,33 +134,33 @@ describe('Backend', () => {
 
         const { locations } = util.filterNodeModules(
             util.mapLocations(
-                (await ctx.backend.references('a', util.createCommit(0), 'src/index.ts', {
+                (await ctx.backend.references(ids.a, 'a', util.createCommit(0), 'src/index.ts', {
                     line: 0,
                     character: 17,
                 })) || { locations: [] }
             )
         )
 
-        expect(locations).toContainEqual(util.createLocation('a', commit, 'src/index.ts', 0, 16, 0, 19)) // def
-        expect(locations).toContainEqual(util.createLocation('a', commit, 'src/index.ts', 11, 18, 11, 21)) // 1st use
-        expect(locations).toContainEqual(util.createLocation('b1', commit, 'src/index.ts', 0, 9, 0, 12)) // import
-        expect(locations).toContainEqual(util.createLocation('b1', commit, 'src/index.ts', 3, 11, 3, 14)) // 1st use
-        expect(locations).toContainEqual(util.createLocation('b2', commit, 'src/index.ts', 0, 9, 0, 12)) // import
-        expect(locations).toContainEqual(util.createLocation('b2', commit, 'src/index.ts', 3, 11, 3, 14)) // 1st use
-        expect(locations).toContainEqual(util.createLocation('b3', commit, 'src/index.ts', 0, 9, 0, 12)) // import
-        expect(locations).toContainEqual(util.createLocation('b3', commit, 'src/index.ts', 3, 11, 3, 14)) // 1st use
-        expect(locations).toContainEqual(util.createLocation('c1', commit, 'src/index.ts', 0, 9, 0, 12)) // import
-        expect(locations).toContainEqual(util.createLocation('c1', commit, 'src/index.ts', 3, 11, 3, 14)) // 1st use
-        expect(locations).toContainEqual(util.createLocation('c1', commit, 'src/index.ts', 3, 15, 3, 18)) // 2nd use
-        expect(locations).toContainEqual(util.createLocation('c1', commit, 'src/index.ts', 3, 26, 3, 29)) // 3rd use
-        expect(locations).toContainEqual(util.createLocation('c2', commit, 'src/index.ts', 0, 9, 0, 12)) // import
-        expect(locations).toContainEqual(util.createLocation('c2', commit, 'src/index.ts', 3, 11, 3, 14)) // 1st use
-        expect(locations).toContainEqual(util.createLocation('c2', commit, 'src/index.ts', 3, 15, 3, 18)) // 2nd use
-        expect(locations).toContainEqual(util.createLocation('c2', commit, 'src/index.ts', 3, 26, 3, 29)) // 3rd use
-        expect(locations).toContainEqual(util.createLocation('c3', commit, 'src/index.ts', 0, 9, 0, 12)) // import
-        expect(locations).toContainEqual(util.createLocation('c3', commit, 'src/index.ts', 3, 11, 3, 14)) // 1st use
-        expect(locations).toContainEqual(util.createLocation('c3', commit, 'src/index.ts', 3, 15, 3, 18)) // 2nd use
-        expect(locations).toContainEqual(util.createLocation('c3', commit, 'src/index.ts', 3, 26, 3, 29)) // 3rd use
+        expect(locations).toContainEqual(util.createLocation(ids.a, commit, 'src/index.ts', 0, 16, 0, 19)) // def
+        expect(locations).toContainEqual(util.createLocation(ids.a, commit, 'src/index.ts', 11, 18, 11, 21)) // 1st use
+        expect(locations).toContainEqual(util.createLocation(ids.b1, commit, 'src/index.ts', 0, 9, 0, 12)) // import
+        expect(locations).toContainEqual(util.createLocation(ids.b1, commit, 'src/index.ts', 3, 11, 3, 14)) // 1st use
+        expect(locations).toContainEqual(util.createLocation(ids.b2, commit, 'src/index.ts', 0, 9, 0, 12)) // import
+        expect(locations).toContainEqual(util.createLocation(ids.b2, commit, 'src/index.ts', 3, 11, 3, 14)) // 1st use
+        expect(locations).toContainEqual(util.createLocation(ids.b3, commit, 'src/index.ts', 0, 9, 0, 12)) // import
+        expect(locations).toContainEqual(util.createLocation(ids.b3, commit, 'src/index.ts', 3, 11, 3, 14)) // 1st use
+        expect(locations).toContainEqual(util.createLocation(ids.c1, commit, 'src/index.ts', 0, 9, 0, 12)) // import
+        expect(locations).toContainEqual(util.createLocation(ids.c1, commit, 'src/index.ts', 3, 11, 3, 14)) // 1st use
+        expect(locations).toContainEqual(util.createLocation(ids.c1, commit, 'src/index.ts', 3, 15, 3, 18)) // 2nd use
+        expect(locations).toContainEqual(util.createLocation(ids.c1, commit, 'src/index.ts', 3, 26, 3, 29)) // 3rd use
+        expect(locations).toContainEqual(util.createLocation(ids.c2, commit, 'src/index.ts', 0, 9, 0, 12)) // import
+        expect(locations).toContainEqual(util.createLocation(ids.c2, commit, 'src/index.ts', 3, 11, 3, 14)) // 1st use
+        expect(locations).toContainEqual(util.createLocation(ids.c2, commit, 'src/index.ts', 3, 15, 3, 18)) // 2nd use
+        expect(locations).toContainEqual(util.createLocation(ids.c2, commit, 'src/index.ts', 3, 26, 3, 29)) // 3rd use
+        expect(locations).toContainEqual(util.createLocation(ids.c3, commit, 'src/index.ts', 0, 9, 0, 12)) // import
+        expect(locations).toContainEqual(util.createLocation(ids.c3, commit, 'src/index.ts', 3, 11, 3, 14)) // 1st use
+        expect(locations).toContainEqual(util.createLocation(ids.c3, commit, 'src/index.ts', 3, 15, 3, 18)) // 2nd use
+        expect(locations).toContainEqual(util.createLocation(ids.c3, commit, 'src/index.ts', 3, 26, 3, 29)) // 3rd use
 
         // Ensure no additional references
         expect(locations?.length).toEqual(20)
@@ -163,33 +173,33 @@ describe('Backend', () => {
 
         const { locations } = util.filterNodeModules(
             util.mapLocations(
-                (await ctx.backend.references('c1', util.createCommit(0), 'src/index.ts', {
+                (await ctx.backend.references(ids.c1, 'c1', util.createCommit(0), 'src/index.ts', {
                     line: 3,
                     character: 16,
                 })) || { locations: [] }
             )
         )
 
-        expect(locations).toContainEqual(util.createLocation('a', commit, 'src/index.ts', 0, 16, 0, 19)) // def
-        expect(locations).toContainEqual(util.createLocation('a', commit, 'src/index.ts', 11, 18, 11, 21)) // 1st use
-        expect(locations).toContainEqual(util.createLocation('b1', commit, 'src/index.ts', 0, 9, 0, 12)) // import
-        expect(locations).toContainEqual(util.createLocation('b1', commit, 'src/index.ts', 3, 11, 3, 14)) // 1st use
-        expect(locations).toContainEqual(util.createLocation('b2', commit, 'src/index.ts', 0, 9, 0, 12)) // import
-        expect(locations).toContainEqual(util.createLocation('b2', commit, 'src/index.ts', 3, 11, 3, 14)) // 1st use
-        expect(locations).toContainEqual(util.createLocation('b3', commit, 'src/index.ts', 0, 9, 0, 12)) // import
-        expect(locations).toContainEqual(util.createLocation('b3', commit, 'src/index.ts', 3, 11, 3, 14)) // 1st use
-        expect(locations).toContainEqual(util.createLocation('c1', commit, 'src/index.ts', 0, 9, 0, 12)) // import
-        expect(locations).toContainEqual(util.createLocation('c1', commit, 'src/index.ts', 3, 11, 3, 14)) // 1st use
-        expect(locations).toContainEqual(util.createLocation('c1', commit, 'src/index.ts', 3, 15, 3, 18)) // 2nd use
-        expect(locations).toContainEqual(util.createLocation('c1', commit, 'src/index.ts', 3, 26, 3, 29)) // 3rd use
-        expect(locations).toContainEqual(util.createLocation('c2', commit, 'src/index.ts', 0, 9, 0, 12)) // import
-        expect(locations).toContainEqual(util.createLocation('c2', commit, 'src/index.ts', 3, 11, 3, 14)) // 1st use
-        expect(locations).toContainEqual(util.createLocation('c2', commit, 'src/index.ts', 3, 15, 3, 18)) // 2nd use
-        expect(locations).toContainEqual(util.createLocation('c2', commit, 'src/index.ts', 3, 26, 3, 29)) // 3rd use
-        expect(locations).toContainEqual(util.createLocation('c3', commit, 'src/index.ts', 0, 9, 0, 12)) // import
-        expect(locations).toContainEqual(util.createLocation('c3', commit, 'src/index.ts', 3, 11, 3, 14)) // 1st use
-        expect(locations).toContainEqual(util.createLocation('c3', commit, 'src/index.ts', 3, 15, 3, 18)) // 2nd use
-        expect(locations).toContainEqual(util.createLocation('c3', commit, 'src/index.ts', 3, 26, 3, 29)) // 3rd use
+        expect(locations).toContainEqual(util.createLocation(ids.a, commit, 'src/index.ts', 0, 16, 0, 19)) // def
+        expect(locations).toContainEqual(util.createLocation(ids.a, commit, 'src/index.ts', 11, 18, 11, 21)) // 1st use
+        expect(locations).toContainEqual(util.createLocation(ids.b1, commit, 'src/index.ts', 0, 9, 0, 12)) // import
+        expect(locations).toContainEqual(util.createLocation(ids.b1, commit, 'src/index.ts', 3, 11, 3, 14)) // 1st use
+        expect(locations).toContainEqual(util.createLocation(ids.b2, commit, 'src/index.ts', 0, 9, 0, 12)) // import
+        expect(locations).toContainEqual(util.createLocation(ids.b2, commit, 'src/index.ts', 3, 11, 3, 14)) // 1st use
+        expect(locations).toContainEqual(util.createLocation(ids.b3, commit, 'src/index.ts', 0, 9, 0, 12)) // import
+        expect(locations).toContainEqual(util.createLocation(ids.b3, commit, 'src/index.ts', 3, 11, 3, 14)) // 1st use
+        expect(locations).toContainEqual(util.createLocation(ids.c1, commit, 'src/index.ts', 0, 9, 0, 12)) // import
+        expect(locations).toContainEqual(util.createLocation(ids.c1, commit, 'src/index.ts', 3, 11, 3, 14)) // 1st use
+        expect(locations).toContainEqual(util.createLocation(ids.c1, commit, 'src/index.ts', 3, 15, 3, 18)) // 2nd use
+        expect(locations).toContainEqual(util.createLocation(ids.c1, commit, 'src/index.ts', 3, 26, 3, 29)) // 3rd use
+        expect(locations).toContainEqual(util.createLocation(ids.c2, commit, 'src/index.ts', 0, 9, 0, 12)) // import
+        expect(locations).toContainEqual(util.createLocation(ids.c2, commit, 'src/index.ts', 3, 11, 3, 14)) // 1st use
+        expect(locations).toContainEqual(util.createLocation(ids.c2, commit, 'src/index.ts', 3, 15, 3, 18)) // 2nd use
+        expect(locations).toContainEqual(util.createLocation(ids.c2, commit, 'src/index.ts', 3, 26, 3, 29)) // 3rd use
+        expect(locations).toContainEqual(util.createLocation(ids.c3, commit, 'src/index.ts', 0, 9, 0, 12)) // import
+        expect(locations).toContainEqual(util.createLocation(ids.c3, commit, 'src/index.ts', 3, 11, 3, 14)) // 1st use
+        expect(locations).toContainEqual(util.createLocation(ids.c3, commit, 'src/index.ts', 3, 15, 3, 18)) // 2nd use
+        expect(locations).toContainEqual(util.createLocation(ids.c3, commit, 'src/index.ts', 3, 26, 3, 29)) // 3rd use
 
         // Ensure no additional references
         expect(locations?.length).toEqual(20)