+++ lsif/src/worker/conversion/conversion.ts
@@ -44,10 +44,10 @@ export async function convertDatabase(
         )
 
         // Move the temp file where it can be found by the server
-        await fs.rename(tempFile, dbFilename(settings.STORAGE_ROOT, upload.id, upload.repository, upload.commit))
+        await fs.rename(tempFile, dbFilename(settings.STORAGE_ROOT, upload.id))
 
         logger.info('Converted upload', {
-            repository: upload.repository,
+            repositoryId: upload.repositoryId,
             commit: upload.commit,
             root: upload.root,
         })
@@ -80,13 +80,21 @@ export async function updateCommitsAndDumpsVisibleFromTip(
 ): Promise<void> {
     const gitserverUrls = fetchConfiguration().gitServers
 
-    const tipCommit = await dumpManager.discoverTip({ repository: upload.repository, gitserverUrls, ctx })
+    const repositoryId = upload.repositoryId
+    const repositoryName = upload.repositoryNameAtUpload
+
+    const tipCommit = await dumpManager.discoverTip({
+        repositoryName,
+        gitserverUrls,
+        ctx,
+    })
     if (tipCommit === undefined) {
         throw new Error('No tip commit available for repository')
     }
 
     const commits = await dumpManager.discoverCommits({
-        repository: upload.repository,
+        repositoryId,
+        repositoryName,
         commit: upload.commit,
         gitserverUrls,
         ctx,
@@ -100,7 +108,8 @@ export async function updateCommitsAndDumpsVisibleFromTip(
         // the tip and all dumps will be invisible.
 
         const tipCommits = await dumpManager.discoverCommits({
-            repository: upload.repository,
+            repositoryId,
+            repositoryName,
             commit: tipCommit,
             gitserverUrls,
             ctx,
@@ -114,6 +123,6 @@ export async function updateCommitsAndDumpsVisibleFromTip(
         }
     }
 
-    await dumpManager.updateCommits(upload.repository, commits, ctx, entityManager)
-    await dumpManager.updateDumpsVisibleFromTip(upload.repository, tipCommit, ctx, entityManager)
+    await dumpManager.updateCommits(upload.repositoryId, commits, ctx, entityManager)
+    await dumpManager.updateDumpsVisibleFromTip(upload.repositoryId, tipCommit, ctx, entityManager)
 }