+++ web/src/search/input/SearchNavbarItem.tsx
@@ -1,13 +1,15 @@
 import * as H from 'history'
-import React, { useCallback } from 'react'
+import React from 'react'
 import { ActivationProps } from '../../../../shared/src/components/activation/Activation'
 import { Form } from '../../components/Form'
 import { submitSearch, QueryState } from '../helpers'
-import { QueryInput } from './QueryInput'
 import { SearchButton } from './SearchButton'
-import { PatternTypeProps, CaseSensitivityProps } from '..'
+import { PatternTypeProps, CaseSensitivityProps, SmartSearchFieldProps } from '..'
+import { LazyMonacoQueryInput } from './LazyMonacoQueryInput'
+import { QueryInput } from './QueryInput'
+import { ThemeProps } from '../../../../shared/src/theme'
 
-interface Props extends ActivationProps, PatternTypeProps, CaseSensitivityProps {
+interface Props extends ActivationProps, PatternTypeProps, CaseSensitivityProps, SmartSearchFieldProps, ThemeProps {
     location: H.Location
     history: H.History
     navbarSearchState: QueryState
@@ -17,45 +19,41 @@ interface Props extends ActivationProps, PatternTypeProps, CaseSensitivityProps 
 /**
  * The search item in the navbar
  */
-export const SearchNavbarItem: React.FunctionComponent<Props> = ({
-    navbarSearchState,
-    onChange,
-    activation,
-    location,
-    history,
-    patternType,
-    caseSensitive,
-    setCaseSensitivity,
-    setPatternType,
-}) => {
-    // Only autofocus the query input on search result pages (otherwise we
-    // capture down-arrow keypresses that the user probably intends to scroll down
-    // in the page).
-    const autoFocus = location.pathname === '/search'
+export class SearchNavbarItem extends React.PureComponent<Props> {
+    private onSubmit = (): void => {
+        const { history, navbarSearchState, patternType, activation, caseSensitive } = this.props
+        submitSearch(history, navbarSearchState.query, 'nav', patternType, caseSensitive, activation)
+    }
 
-    const onSubmit = useCallback(
-        (e: React.FormEvent<HTMLFormElement>): void => {
-            e.preventDefault()
-            submitSearch(history, navbarSearchState.query, 'nav', patternType, caseSensitive, activation)
-        },
-        [history, navbarSearchState.query, patternType, caseSensitive, activation]
-    )
+    private onFormSubmit = (e: React.FormEvent): void => {
+        e.preventDefault()
+        this.onSubmit()
+    }
 
-    return (
-        <Form className="search search--navbar-item d-flex align-items-start flex-grow-1" onSubmit={onSubmit}>
-            <QueryInput
-                value={navbarSearchState}
-                onChange={onChange}
-                autoFocus={autoFocus ? 'cursor-at-end' : undefined}
-                hasGlobalQueryBehavior={true}
-                location={location}
-                history={history}
-                patternType={patternType}
-                setPatternType={setPatternType}
-                caseSensitive={caseSensitive}
-                setCaseSensitivity={setCaseSensitivity}
-            />
-            <SearchButton />
-        </Form>
-    )
+    public render(): React.ReactNode {
+        return (
+            <Form
+                className="search search--navbar-item d-flex align-items-start flex-grow-1"
+                onSubmit={this.onFormSubmit}
+            >
+                {this.props.smartSearchField ? (
+                    <LazyMonacoQueryInput
+                        {...this.props}
+                        hasGlobalQueryBehavior={true}
+                        queryState={this.props.navbarSearchState}
+                        onSubmit={this.onSubmit}
+                        autoFocus={true}
+                    ></LazyMonacoQueryInput>
+                ) : (
+                    <QueryInput
+                        {...this.props}
+                        value={this.props.navbarSearchState}
+                        autoFocus={this.props.location.pathname === '/search' ? 'cursor-at-end' : undefined}
+                        hasGlobalQueryBehavior={true}
+                    ></QueryInput>
+                )}
+                <SearchButton />
+            </Form>
+        )
+    }
 }